+++ src/main/scala/ProbFunctions.scala
@@ -3,7 +3,7 @@ object ProbFunctions {
 
   private def roundToSecondDecimal(floatNum : Float) : Float = (floatNum * 100).round.toFloat / 100.0f
 
-  def moneylineProb(moneyline : Int) : Float = {
+  def moneylineProb(moneyline : Float) : Float = {
     var prob : Float = 0.0f
 
     if (moneyline > 0) {
